fix: Display of the avatar
They updated the hosting location of Trello users’ avatars so that their URI includes the member ID of the user. All of the user’s avatars are now nested under their own member ID. Additionally, they are hosted on a new domain.

Previously, the avatarUrl field on a member looked like this:
https://trello-avatars.s3.amazonaws.com/fc8faaaee46666a4eb8b626c08933e16/170.png 10

Now, the route looks like this:
https://trello-members.s3.amazonaws.com/5589c3ea49b40cedc28cf70e/fc8faaaee46666a4eb8b626c08933e16/170.png 7

Added to 'calendar-for-trello/src/app/models/user.ts' the field: 'avatarUrl' and pasted it into 'calendar-for-trello/src/app/sidebar/sidebar.component.html'